refactor: user page form
